+++ module/model-query/src/internal/types.ts
@@ -1,5 +1,5 @@
-import { DataUtil, FieldConfig } from '@travetto/schema';
-import { Class } from '@travetto/runtime';
+import { FieldConfig, Point } from '@travetto/schema';
+import { Class, toConcrete } from '@travetto/runtime';
 
 const st = (t: string | string[], isArr: boolean = false): Set<string> =>
   new Set((Array.isArray(t) ? t : [t]).map(v => isArr ? `${v}[]` : v));
@@ -16,21 +16,7 @@ const geo = (type: string): Record<string, Set<string>> => ({
   $geoIntersects: st(type, true)
 });
 
-
-export class PointImpl {
-  static validateSchema(input: unknown): 'type' | undefined {
-    const ret = this.bindSchema(input);
-    return ret && !isNaN(ret[0]) && !isNaN(ret[1]) ? undefined : 'type';
-  }
-  static bindSchema(input: unknown): [number, number] | undefined {
-    if (Array.isArray(input) && input.length === 2) {
-      return [
-        DataUtil.coerceType(input[0], Number, false),
-        DataUtil.coerceType(input[1], Number, false)
-      ];
-    }
-  }
-}
+const PointImpl = toConcrete<Point>();
 
 /**
  * Basic type support
