+++ module/schema/src/types.ts
@@ -1,1 +1,8 @@
 export class UnknownType { }
+
+/**
+ * Geometric Point as [number,number] with validation and binding support
+ *
+ * @concrete ../internal/types#PointImpl
+ */
+export type Point = [number, number];
